🐛 printer_safety_check: Fix some issues with new implementation
